+++ src/Tests/DijkstraTest.java
@@ -1,6 +1,8 @@
 package Tests;
 
-import mdsd.*;
+import mdsd.model.DijkstraSolver;
+import mdsd.model.GridElement;
+import mdsd.model.GridManager;
 import org.junit.*;
 import simbad.sim.EnvironmentDescription;
 import simbad.sim.HorizontalWall;
