+++ src/Tests/GoalTest.java
@@ -1,10 +1,11 @@
 package Tests;
 
-import mdsd.*;
+import mdsd.betterproject.BetterAbstractSimulatorMonitor;
+import mdsd.model.*;
+import mdsd.view.SimulatorMonitor;
 import org.junit.*;
 import static org.junit.Assert.*;
 
-import project.AbstractSimulatorMonitor;
 import project.Point;
 import simbad.sim.EnvironmentDescription;
