--- src/mdsd/IRobotController.java
@@ -1,8 +1,0 @@
-package mdsd;
-
-import java.util.List;
-
-public interface IRobotController {
-    public void setMission(int robotId, List<String> missionStr, String strategy);
-
-}
