+++ src/mdsd/controller/IRobotController.java
@@ -3,7 +3,6 @@ package mdsd.controller;
 import java.util.List;
 
 public interface IRobotController {
-    public void setMission(int robotId, List<String> missionStr);
+    public void setMission(int robotId, List<String> missionStr, String strategy);
 
-    //public void addMissionView(IMissionView missionView)
 }
