+++ src/mdsd/model/GridElement.java
@@ -1,4 +1,4 @@
-package mdsd;
+package mdsd.model;
 
 import project.Point;
