+++ src/mdsd/model/IStrategy.java
@@ -1,5 +1,6 @@
-package mdsd;
+package mdsd.model;
 
+import mdsd.model.IGoal;
 import project.Point;
 
 import java.util.Iterator;
