+++ src/mdsd/model/RobotObserver.java
@@ -1,4 +1,6 @@
-package mdsd;
+package mdsd.model;
+
+import mdsd.model.IRobot;
 
 public interface RobotObserver {
     public void update(IRobot robot);
